refactor(utils): consolidate flatted into @vitest/utils/serialization
- Add new serialization module to @vitest/utils with parse, stringify,
  stringifyRaw, stringifyReplace, and toJSON exports
- Migrate all flatted imports to use @vitest/utils/serialization
- Remove duplicate stringifyReplace implementations
- Consistent Error serialization across all usages
