+++ packages/ui/node/reporter.ts
@@ -7,7 +7,7 @@ import { readFile, writeFile } from 'node:fs/promises'
 import { fileURLToPath } from 'node:url'
 import { promisify } from 'node:util'
 import { gzip, constants as zlibConstants } from 'node:zlib'
-import { stringify } from 'flatted'
+import { stringify } from '@vitest/utils/serialization'
 import mime from 'mime/lite'
 import { dirname, extname, relative, resolve } from 'pathe'
 import { globSync } from 'tinyglobby'
